Upgrade React package to v18.2.0 and switch to concurrent mode
The React and React-DOM versions have been updated to 18.2.0, and the ReactDOM.render method has been replaced by ReactDOM.createRoot to switch the application to concurrent mode. This provides a smoother user experience as it allows for the rendering work to be interrupted and resumed, facilitating a more fluid transition between different renders. A 'render' console.log was also added in CharList.js component for debugging and verification purposes.